requirements.txt
# Core
llama-index==0.9.28
llama-index-llms-gemini==0.1.6
chromadb==0.4.18
python-dotenv==1.0.0

# PDF Processing
llama-parse==0.1.4
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0

# Vision
google-generativeai==0.3.1

# GitHub
GitPython==3.1.40
requests==2.31.0

# Web Scraping
trafilatura==1.6.2
beautifulsoup4==4.12.2

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1
pydantic==2.5.0

